# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * JST Tailors Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  username: string;
  /** @nullable */
  email?: string | null;
}

export interface Customer {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface Measurement {
  id: number;
  customerId: number;
  /** @nullable */
  length?: number | null;
  /** @nullable */
  shoulder?: number | null;
  /** @nullable */
  sleeve?: number | null;
  /** @nullable */
  chest?: number | null;
  /** @nullable */
  waist?: number | null;
  /** @nullable */
  hip?: number | null;
  /** @nullable */
  ghera?: number | null;
  /** @nullable */
  shalwarLength?: number | null;
  /** @nullable */
  bottom?: number | null;
  /** @nullable */
  collar?: number | null;
  /** @nullable */
  cuff?: number | null;
  /** @nullable */
  pocket?: number | null;
  /** @nullable */
  frontPatti?: number | null;
  /** @nullable */
  shalwarGhair?: number | null;
  /** @nullable */
  shalwarPocket?: number | null;
  /** @nullable */
  buttonsType?: string | null;
  /** @nullable */
  collarType?: string | null;
  /** @nullable */
  gheraStyle?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  additionalNotes?: string | null;
  updatedAt?: string;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  in_stitching: 'in_stitching',
  ready: 'ready',
  delivered: 'delivered',
} as const;

export interface Order {
  id: number;
  orderNumber: string;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  orderDate: string;
  /** @nullable */
  deliveryDate?: string | null;
  status: OrderStatus;
  totalAmount: number;
  advanceAmount: number;
  balanceAmount: number;
  /** @nullable */
  notes?: string | null;
  createdAt?: string;
  updatedAt?: string;
}

export interface Payment {
  id: number;
  orderId: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  orderNumber?: string | null;
  amount: number;
  paymentDate: string;
  /** @nullable */
  notes?: string | null;
  createdAt?: string;
}

export interface CustomerProfile {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt?: string;
  measurements?: Measurement;
  orders?: Order[];
  payments?: Payment[];
  totalOrders?: number;
  totalPaid?: number;
  totalBalance?: number;
}

export interface CustomerInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
  address?: string;
  notes?: string;
}

export interface CustomerUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  phone?: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface MeasurementInput {
  /** @nullable */
  length?: number | null;
  /** @nullable */
  shoulder?: number | null;
  /** @nullable */
  sleeve?: number | null;
  /** @nullable */
  chest?: number | null;
  /** @nullable */
  waist?: number | null;
  /** @nullable */
  hip?: number | null;
  /** @nullable */
  ghera?: number | null;
  /** @nullable */
  shalwarLength?: number | null;
  /** @nullable */
  bottom?: number | null;
  /** @nullable */
  collar?: number | null;
  /** @nullable */
  cuff?: number | null;
  /** @nullable */
  pocket?: number | null;
  /** @nullable */
  frontPatti?: number | null;
  /** @nullable */
  shalwarGhair?: number | null;
  /** @nullable */
  shalwarPocket?: number | null;
  /** @nullable */
  buttonsType?: string | null;
  /** @nullable */
  collarType?: string | null;
  /** @nullable */
  gheraStyle?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  additionalNotes?: string | null;
}

export type OrderInputStatus = typeof OrderInputStatus[keyof typeof OrderInputStatus];


export const OrderInputStatus = {
  pending: 'pending',
  in_stitching: 'in_stitching',
  ready: 'ready',
  delivered: 'delivered',
} as const;

export interface OrderInput {
  customerId: number;
  orderDate: string;
  /** @nullable */
  deliveryDate?: string | null;
  status?: OrderInputStatus;
  totalAmount: number;
  advanceAmount?: number;
  notes?: string;
}

export type OrderUpdateStatus = typeof OrderUpdateStatus[keyof typeof OrderUpdateStatus];


export const OrderUpdateStatus = {
  pending: 'pending',
  in_stitching: 'in_stitching',
  ready: 'ready',
  delivered: 'delivered',
} as const;

export interface OrderUpdate {
  customerId?: number;
  orderDate?: string;
  /** @nullable */
  deliveryDate?: string | null;
  status?: OrderUpdateStatus;
  totalAmount?: number;
  advanceAmount?: number;
  /** @nullable */
  notes?: string | null;
}

export interface PaymentInput {
  orderId: number;
  customerId: number;
  amount: number;
  paymentDate: string;
  notes?: string;
}

export interface DashboardStats {
  totalCustomers: number;
  totalOrders: number;
  pendingOrders: number;
  inStitchingOrders: number;
  readyOrders: number;
  deliveredOrders: number;
  todayDeliveries: number;
  todayCollection: number;
  remainingBalance: number;
}

export interface DailyReport {
  date: string;
  newOrders: number;
  deliveredOrders: number;
  totalCollection: number;
  newCustomers: number;
  orders?: Order[];
  payments?: Payment[];
}

export type MonthlyReportOrdersByStatus = {
  pending?: number;
  in_stitching?: number;
  ready?: number;
  delivered?: number;
};

export interface MonthlyReport {
  year: number;
  month: number;
  totalOrders: number;
  totalRevenue: number;
  totalCollected: number;
  totalBalance: number;
  newCustomers: number;
  ordersByStatus?: MonthlyReportOrdersByStatus;
}

export interface Settings {
  id: number;
  shopName: string;
  /** @nullable */
  shopLogo?: string | null;
  updatedAt?: string;
}

export interface SettingsUpdate {
  /** @minLength 1 */
  shopName?: string;
  /** @nullable */
  shopLogo?: string | null;
}

export type ListCustomersParams = {
/**
 * Search by name, phone, or address
 */
search?: string;
limit?: number;
offset?: number;
};

export type ListOrdersParams = {
status?: ListOrdersStatus;
customerId?: number;
search?: string;
limit?: number;
offset?: number;
};

export type ListOrdersStatus = typeof ListOrdersStatus[keyof typeof ListOrdersStatus];


export const ListOrdersStatus = {
  pending: 'pending',
  in_stitching: 'in_stitching',
  ready: 'ready',
  delivered: 'delivered',
} as const;

export type ListPaymentsParams = {
customerId?: number;
orderId?: number;
limit?: number;
offset?: number;
};

export type GetDashboardRecentOrdersParams = {
limit?: number;
};

export type GetDailyReportParams = {
/**
 * Date in YYYY-MM-DD format, defaults to today
 */
date?: string;
};

export type GetMonthlyReportParams = {
year?: number;
month?: number;
};