# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: JST Tailors Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers with optional search
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by name, phone, or address
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerProfile"
        "404":
          description: Not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /customers/{id}/measurements:
    get:
      operationId: getCustomerMeasurements
      tags: [measurements]
      summary: Get latest measurements for a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer measurements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Measurement"
    put:
      operationId: upsertCustomerMeasurements
      tags: [measurements]
      summary: Create or update customer measurements
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeasurementInput"
      responses:
        "200":
          description: Upserted measurements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Measurement"
...
  /dashboard/recent-orders:
    get:
      operationId: getDashboardRecentOrders
      tags: [dashboard]
      summary: Get recent orders for dashboard
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username]
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: ["string", "null"]
    Customer:
      type: object
      required: [id, name, phone, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    CustomerProfile:
      type: object
      required: [id, name, phone, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        measurements:
          $ref: "#/components/schemas/Measurement"
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        totalOrders:
          type: integer
        totalPaid:
          type: number
        totalBalance:
          type: number
    CustomerInput:
      type: object
      required: [name, phone]
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        address:
          type: string
        notes:
          type: string
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    Measurement:
      type: object
      required: [id, customerId]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        length:
          type: ["number", "null"]
        shoulder:
          type: ["number", "null"]
        sleeve:
          type: ["number", "null"]
        chest:
          type: ["number", "null"]
        waist:
          type: ["number", "null"]
        hip:
          type: ["number", "null"]
        ghera:
          type: ["number", "null"]
        shalwarLength:
          type: ["number", "null"]
        bottom:
          type: ["number", "null"]
        collar:
          type: ["number", "null"]
        cuff:
          type: ["number", "null"]
        pocket:
          type: ["number", "null"]
        frontPatti:
          type: ["number", "null"]
        shalwarGhair:
          type: ["number", "null"]
        shalwarPocket:
          type: ["number", "null"]
        buttonsType:
          type: ["string", "null"]
        collarType:
          type: ["string", "null"]
        gheraStyle:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        additionalNotes:
          type: ["string", "null"]
        updatedAt:
          type: string
    MeasurementInput:
      type: object
      properties:
        length:
          type: ["number", "null"]
        shoulder:
          type: ["number", "null"]
        sleeve:
          type: ["number", "null"]
        chest:
          type: ["number", "null"]
        waist:
          type: ["number", "null"]
        hip:
          type: ["number", "null"]
        ghera:
          type: ["number", "null"]
        shalwarLength:
          type: ["number", "null"]
        bottom:
          type: ["number", "null"]
        collar:
          type: ["number", "null"]
        cuff:
          type: ["number", "null"]
        pocket:
          type: ["number", "null"]
        frontPatti:
          type: ["number", "null"]
        shalwarGhair:
          type: ["number", "null"]
        shalwarPocket:
          type: ["number", "null"]
        buttonsType:
          type: ["string", "null"]
        collarType:
          type: ["string", "null"]
        gheraStyle:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        additionalNotes:
          type: ["string", "null"]
    Order:
      type: object
      required: [id, orderNumber, customerId, orderDate, status, totalAmount, advanceAmount, balanceAmount]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        orderDate:
          type: string
        deliveryDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_stitching, ready, delivered]
        totalAmount:
          type: number
        advanceAmount:
          type: number
        balanceAmount:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Payment:
      type: object
      required: [id, orderId, customerId, amount, paymentDate]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        orderNumber:
          type: ["string", "null"]
        amount:
          type: number
        paymentDate:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...