# lib/api-zod/src/generated/types/customerUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * JST Tailors Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  phone?: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  notes?: string | null;
}