# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * JST Tailors Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  orderNumber: string;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  orderDate: string;
  /** @nullable */
  deliveryDate?: string | null;
  status: OrderStatus;
  totalAmount: number;
  advanceAmount: number;
  balanceAmount: number;
  /** @nullable */
  notes?: string | null;
  createdAt?: string;
  updatedAt?: string;
}