# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * JST Tailors Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  username: string;
  /** @nullable */
  email?: string | null;
}